style(editor): lint style
